refactor(Game) : 우승자 출력과 관련 기능 Winner 클래스로  분리
